Move the traverse class to Hanami
This allows us to use dependency injection, and not create
many duplicate instances of the same Traverse class
