+++ config/slices/tmas_gate_counts.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module TMASGateCounts
+  class Slice < Hanami::Slice
+    import from: :shared
+  end
+end
