+++ slices/tmas_gate_counts/fetch_tmas_counts.rb
@@ -6,12 +6,12 @@ module TMASGateCounts
   class FetchTMASCounts
     include Dry::Monads[:result]
 
-    include Deps['princeton_timezone', 'tmas_locations', client: 'tmas_client']
+    include Deps['princeton_timezone', 'tmas_locations', 'shared.traverse', client: 'tmas_client']
 
     # Yields a Success([String]) for each day's statistics (or Failure() if there was a problem)
     def call(start_date:, end_date: princeton_timezone.yesterday, locations: tmas_locations.keys)
       (start_date..end_date).each do |date|
-        response = Traverse.new.call(locations) { |location| client.fetch_data(date:, location:) }
+        response = traverse.call(locations) { |location| client.fetch_data(date:, location:) }
         yield response, date
         break if response.failure?
       end
